Hack to allow multiple origin
The code as is assumes same-origin for client/server.  I need to run multiple clients that are not served from the server, but connect to the real-time/websocket pipe.  As it stands there is no way to configure this.

This change is a bit of a hack, and I fully expect you'll likely reject, it, but it does illustrate what I'm looking for.

This allows me to do this in the server's app.js:

```javascript
var client = server.buildClient({host: '192.168.1.10', otherProperty: 'foo'});
fs.writeFileSync('js/client.js', client, 'utf8');
```

Without this hack fix, if I pass in options, they are used to configure the client, but not the transport.  By the time it gets to the transport, you've ignored the options object passed into this generated function, and instead just passed in the `transportConfig` object.